# shared-config/src/main/resources/openapi.yml
openapi: "3.1.0"
info:
  title: "oneblog API"
  description: "oneblog API"
  version: "1.0.0"
servers:
  - url: "https://oneblog.space"
paths:
...
  /api/v1/article/user/{userId}:
    get:
      summary: "GET api/v1/article/user/{userId}"
      operationId: "findArticleByUserId"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ArticleDto"
...
  /api/v1/user:
    post:
      summary: "POST api/v1/user"
      operationId: "saveUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /api/v1/user/nickname/{nickname}:
    get:
      summary: "GET api/v1/user/nickname/{nickname}"
      operationId: "findUserByNickname"
      parameters:
        - name: "nickname"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
  /api/v1/user/{userId}:
    get:
      summary: "GET api/v1/user/{userId}"
      operationId: "findUserByUserId"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
    delete:
      summary: "DELETE api/v1/user/{userId}"
      operationId: "deleteUser"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
  /api/v1/users:
    get:
      summary: "GET api/v1/users"
      operationId: "findAllUsers"
      parameters:
        - name: "page"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int32"
        - name: "size"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
            default: "10"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
components:
  schemas:
...
    LabelDto:
      type: "object"
      properties:
        labelId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
          enum:
            - "Programming_languages"
            - "Java"
            - "CHash"
            - "C"
            - "CPlusPlus"
            - "Python"
            - "JavaScript"
            - "TypeScript"
            - "Sql"
            - "Go"
            - "Kotlin"
            - "Rust"
            - "Php"
            - "Pascal"
            - "Ruby"
            - "Basic"
            - "Assembler"
...
    RoleDto:
      type: "object"
      properties:
        name:
          type: "string"
          enum:
            - "ROLE_USER"
            - "ROLE_ADMIN"
    UserDto:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        nickname:
          type: "string"
        email:
          type: "string"
        roles:
          type: "array"
          items:
            $ref: "#/components/schemas/RoleDto"
...
    PreviewDto:
      type: "object"
      properties:
        articlePreviewId:
          type: "integer"
          format: "int64"
        body:
          type: "string"
    ArticleDto:
      type: "object"
      properties:
        articleId:
          type: "integer"
          format: "int64"
        title:
          type: "string"
        body:
          type: "string"
        createdAt:
          type: "string"
          format: "date-time"
        preview:
          $ref: "#/components/schemas/PreviewDto"
        labels:
          type: "array"
          items:
            $ref: "#/components/schemas/LabelDto"
        user:
          $ref: "#/components/schemas/UserDto"
...
    UserCreateDto:
      type: "object"
      properties:
        name:
          type: "string"
        nickname:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"